added twitter-domains based on https://pastebin.com/BQZq2LpC

+++ domain-generator.sh
@@ -9,19 +9,29 @@ unbound() {
 
         read
 
-        echo "Cole todos os dominios na proxima tela, apos isso utilize ctrl+x para salvar"
-        sleep 3
-        nano /tmp/domain.tmp
-        clear
+        echo "Os dominios do Twitter serão adicionados automaticamente."
+        read -p "Voce quer adicionar mais dominios? (s/n)" resposta
 
-        while read domain;
+        if [[ "$resposta" == "s" ]]; then
+            echo "Cole todos os dominios na proxima tela, apos isso utilize ctrl+x para salvar. Os dominios do Twitter ja serao adicionados por padrao"
+            sleep 3
+            nano /tmp/domain.tmp
+            clear
+
+            while read domain;
+                do
+                    echo local-zone: "$domain" always_refuse
+                done < /tmp/domain.tmp 
+
+            echo ""
+
+            rm -rf /tmp/domain.tmp
+        fi
+
+        while read twitterDomain;
             do
-                echo local-zone: "$domain" always_refuse
-            done < /tmp/domain.tmp 
-
-        echo ""
-
-        rm -rf /tmp/domain.tmp
+                echo local-zone: "$twitterDomain" always_refuse
+            done < ./twitter-domains.txt
 
         echo "Cole todo o acima no arquivo /etc/unbound/blacklist.conf"
         echo "Esse arquivo voce precisará criar!"
@@ -50,32 +60,42 @@ bind9() {
 
     read
 
-    echo "Cole todos os dominios na proxima tela, apos isso utilize ctrl+x para salvar"
-    sleep 3
-    nano /tmp/domain.tmp
-    clear
+    echo "Os dominios do Twitter serão adicionados automaticamente."
+    read -p "Voce quer adicionar mais dominios? (s/n)" resposta
 
-    echo ""
-    echo "$TTL    86400
-@       IN      SOA     localhost. root.localhost. (
-                              1         ; Serial
-                         604800         ; Refresh
-                          86400         ; Retry
-                        2419200         ; Expire
-                          86400 )       ; Negative Cache TTL
-;
-@       IN      NS      localhost.
+    if [[ "$resposta" == "s" ]]; then
+        echo "Cole todos os dominios na proxima tela, apos isso utilize ctrl+x para salvar"
+        sleep 3
+        nano /tmp/domain.tmp
+        clear
 
-"
+        echo ""
+        echo "$TTL    86400
+    @       IN      SOA     localhost. root.localhost. (
+                                1         ; Serial
+                            604800         ; Refresh
+                            86400         ; Retry
+                            2419200         ; Expire
+                            86400 )       ; Negative Cache TTL
+    ;
+    @       IN      NS      localhost.
 
-    while read domain;
-        do
-            echo "$domain     IN CNAME ."
-        done < /tmp/domain.tmp 
+    "
 
-    echo ""
+        while read domain;
+            do
+                echo "$domain     IN CNAME ."
+            done < /tmp/domain.tmp 
 
-    rm -rf /tmp/domain.tmp
+        echo ""
+
+        rm -rf /tmp/domain.tmp
+    fi
+
+    while read twitterDomain;
+    do
+        echo "$twitterDomain     IN CNAME ."
+    done < ./twitter-domains.txt
 
     echo "Cole todo o acima no arquivo db.rpz.zone"
     echo "Esse arquivo voce precisará criar!"
